Change lifelist count based on taxa
Dynamically changes count based on taxa in new lifelist. The count was
originally calculated in the 'species_noapi' component. Now it, along
with the taxa filter labeling logic, are in a shared util function.
