+++ config/locales/en.yml
@@ -5146,6 +5146,7 @@ en:
       list_view: List View
       tree_view: Tree View
       observed_species: Observed Species
+      observed_rank: Observed %{rank}
       unobserved_species: Unobserved Species
       all_species: All Species
       all_unobserved_species: All Unobserved Species
